scripts/requirements.txt
psycopg2-binary
firebase-admin
python-dotenv
